Adding OpenWPM, and changing files
Added OpenWPM, so that now you can run the code anywhere and it will have OpenWPM there.
Continued work on datacollection.py so that it now uses a test URL, and is running the test functions.
Deleted custom_functions.py and put all of its code into datacollection.py so that it can use custom functions properly with OpenWPM.