Continued work on feature E
I did some digging into OpenWPM to find out that they allow for 1 deep crawls. However, we require n deep crawls and also want to see if the URLs found are in iframes or not. I do not think their code looks for links in Iframes, so we will have to combine the current code with theirs (what I pasted in).